feat: enhance inventory service with shipment handling and order validation updates feat: added SNS publisher and published events

+++ inventory-service/src/main/java/com/core/inventoryservice/dto/ItemDTO.java
@@ -7,6 +7,6 @@ import jakarta.validation.constraints.Positive;
 public record ItemDTO(
 		@NotBlank String SKU,
 		@NotBlank String name,
-		@Positive Long quantity,
+		@Positive Integer quantity,
 		@Positive BigDecimal priceAtPurchase
 ) {}
